Organizing example responses
