--- backend/app/api/v1/commons/responses.py
@@ -1,18 +1,0 @@
-def response_200(example):
-    return {
-                "content": {
-                    "application/json": {
-                        "example": example,
-                    }
-                },
-            }
-
-
-def response_422():
-    return {
-                "content": {
-                    "application/json": {
-                        "example": {"error": "invalid date format, start_date must be less than end_date"},
-                    }
-                },
-            }
